We are almost done!
Hey hector,

The project is basically done. All we have to do is implement the
replacement of cards function which is very straightforward. Then all
required tasks for this project is complete. If you do not understand my
code, please contact me. I will be more than happy to explain it!